pyproject.toml
[project]
name = "tekresume-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "fastembed>=0.7.4",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.8",
    "langchain-groq>=1.1.2",
    "langchain-text-splitters>=1.1.0",
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pymupdf>=1.26.7",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
